--- src/main/java/com/example/easybooking/user/SignUpRequest.java
@@ -1,10 +1,0 @@
-package com.example.easybooking.user;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class SignUpRequest {
-    private String nickname;
-}
